Update v2 schemas and implement validation improvements
- Update JSON schemas to use relative $ref paths and improve CollectionID pattern
- Add by_name index schema and update existing index schemas to NDJSON format
- Implement schema validation improvements with better error handling and caching
- Fix test cases to align with updated model properties and schema changes
- Add InternalsVisibleTo attribute for test access to internal members
- Update documentation to reflect schema changes and design decisions

+++ Source/AssetRipper.Tools.AssetDumper.Tests/_TestInfrastructure/Builders/OptionsBuilder.cs
@@ -18,7 +18,7 @@ public class OptionsBuilder
     private string _compression = "none";
     private bool _quiet = true;
     private bool _verbose = false;
-    private long _shardSize = 0; // 0 = auto
+    private int _shardSize = 0; // 0 = auto
     private int _parallelThreads = 1; // Single-threaded for tests by default
     private bool _incrementalMode = false;
     private bool _dryRun = false;
@@ -90,7 +90,7 @@ public class OptionsBuilder
         return this;
     }
 
-    public OptionsBuilder WithShardSize(long size)
+    public OptionsBuilder WithShardSize(int size)
     {
         _shardSize = size;
         return this;
